Split layer tests into their own files.
