+++ src/mlpack/tests/CMakeLists.txt
@@ -6,7 +6,8 @@ add_executable(mlpack_test
   # Tests for neural network code (and reinforcement learning code).
   ann/activation_functions_test.cpp
   ann/ann_dist_test.cpp
-  ann/ann_layer_test.cpp
+  # This includes all files in ann/layer/.
+  ann/layer_test.cpp
   ann/ann_regularizer_test.cpp
   ann/ann_test_tools.hpp
   ann/convolutional_network_test.cpp
